+++ modules/sor/sorV2/sor-port/pools/fx/fxPool.ts
@@ -3,8 +3,8 @@ import { BasePool, SwapKind } from '../../types';
 import { GqlPoolType } from '../../../../../../schema';
 import { FxPoolToken } from './fxPoolToken';
 import { PrismaPoolWithDynamic } from '../../../../../../prisma/prisma-types';
-import { Token } from '../../token';
-import { TokenAmount } from '../../tokenAmount';
+import { Token } from '../../entities/token';
+import { TokenAmount } from '../../entities/tokenAmount';
 import { MathFx, parseFixedCurveParam } from './helpers';
 import { FxData } from '../../../../../pool/subgraph-mapper';
 import { Chain } from '@prisma/client';
