+++ modules/sor/sorV2/sor-port/utils/helpers.ts
@@ -1,5 +1,5 @@
-import { Token } from '../token';
-import { BigintIsh, TokenAmount } from '../tokenAmount';
+import { Token } from '../entities/token';
+import { BigintIsh, TokenAmount } from '../entities/tokenAmount';
 import { SwapKind } from '../types';
 
 export function checkInputs(
