+++ modules/sor/sorV2/sorV2.service.ts
@@ -9,12 +9,12 @@ import { poolsToIgnore } from '../constants';
 import { AllNetworkConfigsKeyedOnChain } from '../../network/network-config';
 import * as Sentry from '@sentry/node';
 import { getToken } from '../utils';
-import { Swap } from './sor-port/swap';
+import { Swap } from './sor-port/entities/swap';
 import { Address, parseUnits } from 'viem';
 import { BatchSwapStep, SingleSwap, SwapKind } from './sor-port/types';
 import { sorGetSwapsWithPools } from './sor-port/static';
 import { SwapResultV2 } from './swapResultV2';
-import { TokenAmount } from './sor-port/tokenAmount';
+import { TokenAmount } from './sor-port/entities/tokenAmount';
 import { poolService } from '../../pool/pool.service';
 import { mapRoutes } from './beetsHelpers';
 import { replaceZeroAddressWithEth } from '../../web3/addresses';
